Add unit tests for ValidationFailure extensions.
