+++ src/back-end/TodoList.Application.Tests/TodoItems/TodoItemResultTests.cs
@@ -6,6 +6,7 @@ using TodoList.Application.TodoItems;
 
 namespace TodoList.Application.Tests.TodoItems
 {
+    [ExcludeFromCodeCoverage(Justification = "Tests")]
     public class TodoItemResultTests
     {
         [Fact]
