+++ src/back-end/TodoList.Application.Tests/TodoList.Application.Tests.csproj
@@ -29,5 +29,9 @@
   <ItemGroup>
     <Using Include="Xunit" />
   </ItemGroup>
+  
+  <ItemGroup>
+    <Folder Include="Common\Extensions\ValidationExtensions\" />
+  </ItemGroup>
 
 </Project>
